rename typing.py to _typing.py
