+++ src/spei/_typing.py
